+++ depthcharge/models/__init__.py
@@ -1,2 +1,0 @@
-"""End-to-end models and dataloaders."""
-from .denovo import Spec2Pep